requirements.txt
customtkinter
pymodbus
pyserial
